spider_manager/requirements.txt
../src/crawling_utils/
../src/step-by-step/
../src/captcha_solver/
../src/scrapy_puppeteer/
../src/cssify/
attrs==18.1.0  # Updated from 17.2.0
cffi==1.11.5  # Updated from 1.10.0
ConcurrentLogHandler==0.9.1
cryptography==2.3
cssselect==1.0.3  # Updated from 1.0.1
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
idna==2.6
ipaddress==1.0.22  # Updated from 1.0.18
kafka-python==1.4.3  # Updated from 1.3.4
kazoo==2.4.0
lxml==4.2.1  # Updated from 3.8.0
mock==2.0.0
nose==1.3.7
parsel==1.4.0  # Updated from 1.2.0
pbr==4.0.3  # Updated from 3.1.1
pyasn1==0.4.3  # Updated from 0.3.2
pyasn1-modules==0.2.1  # Updated from 0.0.11
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0  # Updated from 17.2.0
python-json-logger==0.1.8
PyYAML==3.12
queuelib==1.5.0  # Updated from 1.4.2
redis>=3.0  # Updated from 2.10.5
requests==2.18.4  # Updated from 2.18.3
requests-file==1.4.3  # Updated from 1.4.2
retrying==1.3.3
Scrapy==1.5.0  # Updated from 1.4.0
scutils/
service-identity==17.0.0
six==1.11.0  # Updated from 1.10.0
testfixtures==6.0.2  # Updated from 5.1.1
tldextract==2.2.0  # Updated from 2.1.0
Twisted==18.4.0  # Updated from 17.5.0
ujson==1.35
w3lib==1.19.0  # Updated from 1.18.0
zope.interface==4.5.0  # Updated from 4.4.2